+++ src/main/java/org/agilewiki/utils/lazydurable/LazyDurableListNode.java
@@ -19,7 +19,7 @@ public class LazyDurableListNode {
     /**
      * A time after all insertions and deletions.
      */
-    public final static long MAX_TIME = Integer.MAX_VALUE - 1;
+    public final static long MAX_TIME = Long.MAX_VALUE - 1;
 
     /**
      * The root node of an empty tree.
@@ -418,7 +418,7 @@ public class LazyDurableListNode {
      * @return The revised root node.
      */
     public LazyDurableListNode add(int ndx, Object value, long time) {
-        return add(ndx, value, time, Integer.MAX_VALUE);
+        return add(ndx, value, time, Long.MAX_VALUE);
     }
 
     protected LazyDurableListNode add(int ndx, Object value, long created, long deleted) {
