+++ src/main/java/org/agilewiki/utils/lazydurable/LazyDurableMapNode.java
@@ -123,7 +123,7 @@ public class LazyDurableMapNode {
      * @return The revised root node.
      */
     public LazyDurableMapNode add(Comparable key, int ndx, Object value, long time) {
-        return add(key, ndx, value, time, Integer.MAX_VALUE);
+        return add(key, ndx, value, time, Long.MAX_VALUE);
     }
 
     protected LazyDurableMapNode add(Comparable key, int ndx, Object value, long created, long deleted) {
