+++ src/main/java/org/agilewiki/utils/maplist/MapNode.java
@@ -139,7 +139,7 @@ public class MapNode {
      * @return The revised root node.
      */
     public MapNode add(Comparable key, int ndx, Object value, long time) {
-        return add(key, ndx, value, time, Integer.MAX_VALUE);
+        return add(key, ndx, value, time, Long.MAX_VALUE);
     }
 
     protected MapNode add(Comparable key, int ndx, Object value, long created, long deleted) {
